requirements.txt
selenium>=4.15.0
beautifulsoup4>=4.12.0
webdriver-manager>=4.0.0
pyyaml>=6.0.0
psycopg2-binary>=2.9.0